Match architecture to `construct_circuits()`
